Tried a differnt way to place cabins on map. changed to use % instead of pixels, changed position to type double, also changed currency to EURO in booking confirmation.

+++ src/HejCamping.Web/Models/CabinViewModel.cs
@@ -5,8 +5,8 @@ namespace HejCamping.Web.Models
         public int Id { get; set; }
         public int Number { get; set; }
         public bool IsVacant { get; set; }
-        public int PositionX { get; set; } 
-        public int PositionY { get; set; } 
+        public double PositionX { get; set; } 
+        public double PositionY { get; set; } 
         public int PricePerNight { get; set; }
     }
 }
